requirements.txt
annotated-types==0.6.0
arrow==1.3.0
attrs==23.1.0
black==23.11.0
blinker==1.6.2
bravado==11.0.3
bravado-core==6.1.1
bravado-types==1.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Faker==21.0.0
flake8==6.1.0
Flask==3.0.0
fqdn==1.5.1
idna==3.6
iniconfig==2.0.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsf==0.9.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jsonschema-typed-v2==0.8.0
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
monotonic==1.6
msgpack==1.0.7
mypy==1.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
pytest==7.4.3
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rstr==3.2.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
swagger-spec-validator==3.0.3
tomli==2.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==2.1.0
webcolors==1.13
Werkzeug==3.0.1